% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{make_agenda}
\alias{make_agenda}
\title{Return a character vector of markdown content for the agenda slide}
\usage{
make_agenda(repo)
}
\description{
Return a character vector of markdown content for the agenda slide
}
