% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{make_md}
\alias{make_md}
\title{Generates a markdown slide deck from a Carpentries workbench project}
\usage{
make_md(repo, verbose = FALSE)
}
\arguments{
\item{repo}{A length-1 character vector pointing to the root of a Carpentries workbench project}

\item{verbose}{Logical scalar. TRUE if additional but non-essential logging should be provided.}
}
\description{
Generates a markdown slide deck from a Carpentries workbench project
}
