% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reveal.R
\docType{data}
\name{post_reveal}
\alias{post_reveal}
\title{Path to the reveal JS postprocessing lua filter}
\format{
An object of class \code{character} of length 1.
}
\usage{
post_reveal
}
\description{
Path to the reveal JS postprocessing lua filter
}
\keyword{datasets}
